refactor(wave2-step2): mechanically split runtime impl behind facades
